+++ apps/tetris-classic/src/features/game/board/utils/mergeFigure.ts
@@ -8,10 +8,10 @@ export function mergeFigure(board: Matrix, figure: Figure): Matrix {
     const [figureWidth, figureHeight] = getMatrixSize(figure.figure);
     for (let y = 0; y < figureHeight; y++) {
         for (let x = 0; x < figureWidth; x++) {
-            let figureCell = figure.figure[y][x];
+            const figureCell = figure.figure[y][x];
             if (figureCell) {
-                let boardY = figure.position.y + y;
-                let boardX = figure.position.x + x;
+                const boardY = figure.position.y + y;
+                const boardX = figure.position.x + x;
                 merge[boardY][boardX] = figureCell;
             }
         }