+++ apps/tetris-classic/src/utils/deferred.ts
@@ -2,14 +2,14 @@ import { notNull } from './typescript';
 
 export type Deferred<T> = {
     resolve: (value: T) => void;
-    reject: (reason?: any) => void;
+    reject: (reason?: unknown) => void;
     promise: Promise<T>;
 };
 
 export function createDeferred<T>(): Deferred<T> {
-    var resolve: ((value: T) => void) | null = null;
-    var reject: ((reason?: any) => void) | null = null;
-    var promise = new Promise<T>((_resolve, _reject) => {
+    let resolve: ((value: T) => void) | null = null;
+    let reject: ((reason?: unknown) => void) | null = null;
+    const promise = new Promise<T>((_resolve, _reject) => {
         resolve = _resolve;
         reject = _reject;
     });