+++ apps/tetris-classic/src/utils/typescript.ts
@@ -6,9 +6,3 @@ export function notNull<T>(value: T | null, name: string): T | never {
     }
     return value;
 }
-
-export const getObjectKeys = Object.keys as <T extends object>(obj: T) => Array<keyof T>;
-
-export type Fn = (...args: any[]) => any;
-
-export type Mutable<T> = { -readonly [P in keyof T]: T[P] };