+++ packages/utils/redux-subslice/src/index.ts
@@ -1,0 +1,6 @@
+import { declareSubslice } from './subslice/declareSubslice';
+import { createSubslices } from './subslice/createSubslices';
+
+export { declareSubslice, createSubslices };
+
+export * from './types';