+++ packages/utils/typescript/src/index.ts
@@ -1,0 +1,5 @@
+export const getObjectKeys = Object.keys as <T extends object>(obj: T) => Array<keyof T>;
+
+export type Fn = (...args: any[]) => any;
+
+export type Mutable<T> = { -readonly [P in keyof T]: T[P] };