+++ apps/tetris-classic/CHANGELOG.md
@@ -1,5 +1,16 @@
 # tetris-classic
 
+## 0.3.0
+
+### Minor Changes
+
+- Extract utils packages (typescript and redux-subslice)
+
+### Patch Changes
+
+- Updated dependencies
+  - @muravjev/utils-redux-subslice@0.1.0
+
 ## 0.2.0
 
 ### Minor Changes